Refactor integrationTestHelper
