+++ test/integration/user_downvotes_page.test.js
@@ -1,5 +1,5 @@
 const assert = require("assert");
-const { getExtensionId } = require('../integrationTestHelper');
+const { getExtensionId } = require('../test_helper');
 
 describe('User downvotes page', function () {
   let page;
