+++ test/integration/user_upvotes_page.test.js
@@ -1,5 +1,5 @@
 const assert = require("assert");
-const { getExtensionId } = require('../integrationTestHelper');
+const { getExtensionId } = require('../test_helper');
 
 describe('User upvotes page', () => {
   let page;
